refactor: migrate numeric types from float to double across ADTs
- Update Ponto, Circulo and Matriz interfaces and implementations
- Improve numerical precision for geometric and matrix operations
- Adjust main tests accordingly

+++ include/circulo.h
@@ -19,13 +19,13 @@ typedef struct circulo_t Circulo;
  * Aloca e retorna um circulo no plano cartesiano, com centro em (x, y) e raio r
  *
  * Parametros:
- * float x: abcissa do centro do circulo
- * float y: ordenada do centro do circulo
- * float r: raio do circulo
+ * double x: abcissa do centro do circulo
+ * double y: ordenada do centro do circulo
+ * double r: raio do circulo
  *
  * Retorna Circulo*: circulo com centro em (x, y) e raio r
  */
-Circulo* circulo_cria(float x, float y, float r);
+Circulo* circulo_cria(double x, double y, double r);
 
 /**
  * Libera a memoria de um circulo previamente criado
@@ -40,22 +40,22 @@ void circulo_libera(Circulo *c);
  *
  * Parametros:
  * Circulo* c: ponteiro para um circulo
- * float* x: armazena abcissa do centro do circulo
- * float* y: armazena ordenada do centro do circulo
- * float* r: armazena o raio do circulo
+ * double* x: armazena abcissa do centro do circulo
+ * double* y: armazena ordenada do centro do circulo
+ * double* r: armazena o raio do circulo
  */
-void circulo_acessa(Circulo* c, float* x, float* y, float* r);
+void circulo_acessa(Circulo* c, double* x, double* y, double* r);
 
 /**
  * Atribui novos valores as coordenadas do centro e ao raio de um circulo
  *
  * Parametros:
  * Circulo* c: ponteiro para um circulo
- * float x: nova abcissa do centro do circulo
- * float y: nova ordenada do centro do circulo
- * float r: novo raio do circulo
+ * double x: nova abcissa do centro do circulo
+ * double y: nova ordenada do centro do circulo
+ * double r: novo raio do circulo
  */
-void circulo_atribui(Circulo* c, float x, float y, float r);
+void circulo_atribui(Circulo* c, double x, double y, double r);
 
 /**
  * Retorna a area de um circulo
@@ -63,9 +63,9 @@ void circulo_atribui(Circulo* c, float x, float y, float r);
  * Parametros:
  * Circulo* c: ponteiro para um circulo
  *
- * Retorna float: a area do circulo
+ * Retorna double: a area do circulo
  */
-float circulo_area(Circulo* c);
+double circulo_area(Circulo* c);
 
 /**
  * Retorna se um ponto pertence a um circulo
